disable PSQL

+++ scripts/utils/psql_connector.py
@@ -16,9 +16,15 @@ class PSQLConnector:
         self.port = os.getenv("DB_PORT")
 
     def connect(self):
+        # run only if DB is configured
+        if self.dbname is None:
+            return
         self.engine = create_engine(f'postgresql://{self.user}:{self.password}@{self.host}:{self.port}/{self.dbname}')
 
     def drop_table_if_exists(self, table_name):
+        # run only if DB is configured
+        if self.dbname is None:
+            return
         try:
             with self.engine.connect() as conn:
                 conn.execute(f"DROP TABLE IF EXISTS {table_name}")
@@ -27,6 +33,9 @@ class PSQLConnector:
             self.logger.error(f"An error occurred while dropping the table: {e}")
 
     def save_df_to_sql(self, df, table_name, chunksize=1000, if_exists='append', index=False):
+        # run only if DB is configured
+        if self.dbname is None:
+            return
         try:
             self.drop_table_if_exists(table_name)
             df.to_sql(table_name, self.engine, if_exists=if_exists, index=index, chunksize=chunksize)
